Added interface for report service
